pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aritech-ats"
version = "0.2.0"
description = "Unofficial Python client library for ATS alarm panels"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Seba Kerckhof"}
]
keywords = ["aritech", "ats", "alarm", "security", "home-automation", "carrier"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
    "Topic :: Security",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]
dependencies = [
    "pycryptodome>=3.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.7.0",
]

[project.scripts]
aritech-cli = "aritech_ats.cli:main"

[project.urls]
Homepage = "https://github.com/sebakerckhof/atsadvisor"
Repository = "https://github.com/sebakerckhof/atsadvisor"

[tool.hatch.build.targets.wheel]
packages = ["src/aritech_ats"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "ASYNC"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["aritech_ats"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
